Get or reset headers for one or all header field(s) respectively

Get the expected header(s) for one field or reset all header fields

Usage

get_header(k = NULL)

Arguments

k

string of field (data type) to return headers for

Value

For get_header a character vector of headers for field k.

Details

If get_header is called with no values, the entire available header list is returned.

Examples

get_header(k = NULL)
#> $manifest
#> $manifest$barcode
#> [1] "Barcode" "Plate"  
#> 
#> $manifest$template
#> [1] "Template"  "Treatment"
#> 
#> $manifest$duration
#> [1] "Duration"
#> 
#> 
#> $raw_data
#> [1] "ReadoutValue"    "BackgroundValue" "UntrtReadout"    "Day0Readout"    
#> [5] "masked"         
#> 
#> $normalized_results
#> [1] "x"                    "CorrectedReadout"     "GRvalue"             
#> [4] "RelativeViability"    "DivisionTime"         "RefGRvalue"          
#> [7] "RefRelativeViability"
#> 
#> $averaged_results
#> [1] "x"                     "x_std"                 "std_RelativeViability"
#> [4] "std_GRvalue"          
#> 
#> $response_metrics
#>  [1] "x_mean"      "x_AOC"       "x_AOC_range" "xc50"        "x_max"      
#>  [6] "ec50"        "x_inf"       "x_0"         "h"           "r2"         
#> [11] "p_value"     "rss"         "x_sd_avg"    "fit_type"   
#> 
#> $metric_average_fields
#> $metric_average_fields$mean
#> [1] "x_mean"      "x_AOC"       "x_AOC_range" "x_max"       "x_inf"      
#> [6] "x_0"        
#> 
#> $metric_average_fields$geometric_mean
#>  [1] "xc50"    "ec50"    "GR50"    "GEC50"   "IC50"    "EC50"    "GR_xc50"
#>  [8] "RV_xc50" "GR_ec50" "RV_ec50"
#> 
#> $metric_average_fields$fit_type
#> [1] "fit_type"    "Fit Type"    "Fit Type RV" "Fit Type GR" "RV_fit_type"
#> [6] "GR_fit_type"
#> 
#> $metric_average_fields$blacklisted
#> [1] "Tissue"                       "cellline_tissue"             
#> [3] "Reference Division Time"      "cellline_ref_div_time"       
#> [5] "Parental Identifier"          "cellline_parental_identifier"
#> 
#> 
#> $add_clid
#> $add_clid$cellline_name
#> [1] "CellLineName"
#> 
#> $add_clid$cellline_tissue
#> [1] "Tissue"
#> 
#> $add_clid$cellline_parental_identifier
#> [1] "parental_identifier"
#> 
#> $add_clid$cellline_subtype
#> [1] "subtype"
#> 
#> $add_clid$cellline_ref_div_time
#> [1] "ReferenceDivisionTime"
#> 
#> 
#> $metrics_names
#>    x_mean    x_AOC    x_AOC_range    xc50   x_max    ec50    x_inf    x_0   
#> RV "RV_mean" "RV_AOC" "RV_AOC_range" "IC50" "E_max"  "EC50"  "E_inf"  "E_0" 
#> GR "GR_mean" "GR_AOC" "GR_AOC_range" "GR50" "GR_max" "GEC50" "GR_inf" "GR_0"
#>    h      r2      p_value      rss      x_sd_avg    fit_type     
#> RV "h_RV" "RV_r2" "RV_p_value" "RV_rss" "RV_sd_avg" "fit_type_RV"
#> GR "h_GR" "GR_r2" "GR_p_value" "GR_rss" "GR_sd_avg" "fit_type_GR"
#> 
#> $metrics_results
#>  [1] "maxlog10Concentration" "N_conc"                "cotrt_value"          
#>  [4] "source"                "x_mean"                "x_AOC"                
#>  [7] "x_AOC_range"           "xc50"                  "x_max"                
#> [10] "ec50"                  "x_inf"                 "x_0"                  
#> [13] "h"                     "r2"                    "p_value"              
#> [16] "rss"                   "x_sd_avg"              "fit_type"             
#> [19] "RV_mean"               "GR_mean"               "RV_AOC"               
#> [22] "GR_AOC"                "RV_AOC_range"          "GR_AOC_range"         
#> [25] "IC50"                  "GR50"                  "E_max"                
#> [28] "GR_max"                "EC50"                  "GEC50"                
#> [31] "E_inf"                 "GR_inf"                "E_0"                  
#> [34] "GR_0"                  "h_RV"                  "h_GR"                 
#> [37] "RV_r2"                 "GR_r2"                 "RV_p_value"           
#> [40] "GR_p_value"            "RV_rss"                "GR_rss"               
#> [43] "RV_sd_avg"             "GR_sd_avg"             "fit_type_RV"          
#> [46] "fit_type_GR"          
#> 
#> $controlled
#> $controlled[[1]]
#> [1] "clid"
#> 
#> $controlled$barcode
#> [1] "Barcode" "Plate"  
#> 
#> $controlled$template
#> [1] "Template"  "Treatment"
#> 
#> $controlled$duration
#> [1] "Duration"
#> 
#> $controlled[[5]]
#> [1] "Gnumber"
#> 
#> $controlled[[6]]
#> [1] "Concentration"
#> 
#> $controlled[[7]]
#> [1] "Gnumber_2"
#> 
#> $controlled[[8]]
#> [1] "Gnumber_3"
#> 
#> $controlled[[9]]
#> [1] "Gnumber_4"
#> 
#> $controlled[[10]]
#> [1] "Gnumber_5"
#> 
#> $controlled[[11]]
#> [1] "Gnumber_6"
#> 
#> $controlled[[12]]
#> [1] "Gnumber_7"
#> 
#> $controlled[[13]]
#> [1] "Gnumber_8"
#> 
#> $controlled[[14]]
#> [1] "Gnumber_9"
#> 
#> $controlled[[15]]
#> [1] "Gnumber_10"
#> 
#> $controlled[[16]]
#> [1] "Concentration_2"
#> 
#> $controlled[[17]]
#> [1] "Concentration_3"
#> 
#> $controlled[[18]]
#> [1] "Concentration_4"
#> 
#> $controlled[[19]]
#> [1] "Concentration_5"
#> 
#> $controlled[[20]]
#> [1] "Concentration_6"
#> 
#> $controlled[[21]]
#> [1] "Concentration_7"
#> 
#> $controlled[[22]]
#> [1] "Concentration_8"
#> 
#> $controlled[[23]]
#> [1] "Concentration_9"
#> 
#> $controlled[[24]]
#> [1] "Concentration_10"
#> 
#> 
#> $reserved
#> $reserved$cellline_name
#> [1] "CellLineName"
#> 
#> $reserved$cellline_tissue
#> [1] "Tissue"
#> 
#> $reserved$cellline_parental_identifier
#> [1] "parental_identifier"
#> 
#> $reserved$cellline_subtype
#> [1] "subtype"
#> 
#> $reserved$cellline_ref_div_time
#> [1] "ReferenceDivisionTime"
#> 
#> $reserved[[6]]
#> [1] "DrugName"
#> 
#> $reserved[[7]]
#> [1] "masked"
#> 
#> $reserved[[8]]
#> [1] "DrugName_2"
#> 
#> $reserved[[9]]
#> [1] "DrugName_3"
#> 
#> $reserved[[10]]
#> [1] "DrugName_4"
#> 
#> $reserved[[11]]
#> [1] "DrugName_5"
#> 
#> $reserved[[12]]
#> [1] "DrugName_6"
#> 
#> $reserved[[13]]
#> [1] "DrugName_7"
#> 
#> $reserved[[14]]
#> [1] "DrugName_8"
#> 
#> $reserved[[15]]
#> [1] "DrugName_9"
#> 
#> $reserved[[16]]
#> [1] "DrugName_10"
#> 
#> $reserved[[17]]
#> [1] "drug_moa_2"
#> 
#> $reserved[[18]]
#> [1] "drug_moa_3"
#> 
#> $reserved[[19]]
#> [1] "drug_moa_4"
#> 
#> $reserved[[20]]
#> [1] "drug_moa_5"
#> 
#> $reserved[[21]]
#> [1] "drug_moa_6"
#> 
#> $reserved[[22]]
#> [1] "drug_moa_7"
#> 
#> $reserved[[23]]
#> [1] "drug_moa_8"
#> 
#> $reserved[[24]]
#> [1] "drug_moa_9"
#> 
#> $reserved[[25]]
#> [1] "drug_moa_10"
#> 
#> $reserved[[26]]
#> [1] "ReadoutValue"
#> 
#> $reserved[[27]]
#> [1] "BackgroundValue"
#> 
#> $reserved[[28]]
#> [1] "UntrtReadout"
#> 
#> $reserved[[29]]
#> [1] "Day0Readout"
#> 
#> $reserved[[30]]
#> [1] "masked"
#> 
#> $reserved[[31]]
#> [1] "x"
#> 
#> $reserved[[32]]
#> [1] "CorrectedReadout"
#> 
#> $reserved[[33]]
#> [1] "GRvalue"
#> 
#> $reserved[[34]]
#> [1] "RelativeViability"
#> 
#> $reserved[[35]]
#> [1] "DivisionTime"
#> 
#> $reserved[[36]]
#> [1] "RefGRvalue"
#> 
#> $reserved[[37]]
#> [1] "RefRelativeViability"
#> 
#> $reserved[[38]]
#> [1] "x"
#> 
#> $reserved[[39]]
#> [1] "x_std"
#> 
#> $reserved[[40]]
#> [1] "std_RelativeViability"
#> 
#> $reserved[[41]]
#> [1] "std_GRvalue"
#> 
#> $reserved[[42]]
#> [1] "maxlog10Concentration"
#> 
#> $reserved[[43]]
#> [1] "N_conc"
#> 
#> $reserved[[44]]
#> [1] "cotrt_value"
#> 
#> $reserved[[45]]
#> [1] "source"
#> 
#> $reserved[[46]]
#> [1] "x_mean"
#> 
#> $reserved[[47]]
#> [1] "x_AOC"
#> 
#> $reserved[[48]]
#> [1] "x_AOC_range"
#> 
#> $reserved[[49]]
#> [1] "xc50"
#> 
#> $reserved[[50]]
#> [1] "x_max"
#> 
#> $reserved[[51]]
#> [1] "ec50"
#> 
#> $reserved[[52]]
#> [1] "x_inf"
#> 
#> $reserved[[53]]
#> [1] "x_0"
#> 
#> $reserved[[54]]
#> [1] "h"
#> 
#> $reserved[[55]]
#> [1] "r2"
#> 
#> $reserved[[56]]
#> [1] "p_value"
#> 
#> $reserved[[57]]
#> [1] "rss"
#> 
#> $reserved[[58]]
#> [1] "x_sd_avg"
#> 
#> $reserved[[59]]
#> [1] "fit_type"
#> 
#> $reserved[[60]]
#> [1] "RV_mean"
#> 
#> $reserved[[61]]
#> [1] "GR_mean"
#> 
#> $reserved[[62]]
#> [1] "RV_AOC"
#> 
#> $reserved[[63]]
#> [1] "GR_AOC"
#> 
#> $reserved[[64]]
#> [1] "RV_AOC_range"
#> 
#> $reserved[[65]]
#> [1] "GR_AOC_range"
#> 
#> $reserved[[66]]
#> [1] "IC50"
#> 
#> $reserved[[67]]
#> [1] "GR50"
#> 
#> $reserved[[68]]
#> [1] "E_max"
#> 
#> $reserved[[69]]
#> [1] "GR_max"
#> 
#> $reserved[[70]]
#> [1] "EC50"
#> 
#> $reserved[[71]]
#> [1] "GEC50"
#> 
#> $reserved[[72]]
#> [1] "E_inf"
#> 
#> $reserved[[73]]
#> [1] "GR_inf"
#> 
#> $reserved[[74]]
#> [1] "E_0"
#> 
#> $reserved[[75]]
#> [1] "GR_0"
#> 
#> $reserved[[76]]
#> [1] "h_RV"
#> 
#> $reserved[[77]]
#> [1] "h_GR"
#> 
#> $reserved[[78]]
#> [1] "RV_r2"
#> 
#> $reserved[[79]]
#> [1] "GR_r2"
#> 
#> $reserved[[80]]
#> [1] "RV_p_value"
#> 
#> $reserved[[81]]
#> [1] "GR_p_value"
#> 
#> $reserved[[82]]
#> [1] "RV_rss"
#> 
#> $reserved[[83]]
#> [1] "GR_rss"
#> 
#> $reserved[[84]]
#> [1] "RV_sd_avg"
#> 
#> $reserved[[85]]
#> [1] "GR_sd_avg"
#> 
#> $reserved[[86]]
#> [1] "fit_type_RV"
#> 
#> $reserved[[87]]
#> [1] "fit_type_GR"
#> 
#> $reserved[[88]]
#> [1] "WellRow"
#> 
#> $reserved[[89]]
#> [1] "WellColumn"
#> 
#> 
#> $ordered_1
#>  [1] "CellLineName"     "Tissue"           "Duration"         "DrugName"        
#>  [5] "Concentration"    "DrugName_2"       "Concentration_2"  "DrugName_3"      
#>  [9] "Concentration_3"  "DrugName_4"       "Concentration_4"  "DrugName_5"      
#> [13] "Concentration_5"  "DrugName_6"       "Concentration_6"  "DrugName_7"      
#> [17] "Concentration_7"  "DrugName_8"       "Concentration_8"  "DrugName_9"      
#> [21] "Concentration_9"  "DrugName_10"      "Concentration_10"
#> 
#> $ordered_2
#> $ordered_2[[1]]
#> [1] "x"
#> 
#> $ordered_2[[2]]
#> [1] "CorrectedReadout"
#> 
#> $ordered_2[[3]]
#> [1] "GRvalue"
#> 
#> $ordered_2[[4]]
#> [1] "RelativeViability"
#> 
#> $ordered_2[[5]]
#> [1] "DivisionTime"
#> 
#> $ordered_2[[6]]
#> [1] "RefGRvalue"
#> 
#> $ordered_2[[7]]
#> [1] "RefRelativeViability"
#> 
#> $ordered_2[[8]]
#> [1] "x"
#> 
#> $ordered_2[[9]]
#> [1] "x_std"
#> 
#> $ordered_2[[10]]
#> [1] "std_RelativeViability"
#> 
#> $ordered_2[[11]]
#> [1] "std_GRvalue"
#> 
#> $ordered_2[[12]]
#> [1] "maxlog10Concentration"
#> 
#> $ordered_2[[13]]
#> [1] "N_conc"
#> 
#> $ordered_2[[14]]
#> [1] "cotrt_value"
#> 
#> $ordered_2[[15]]
#> [1] "source"
#> 
#> $ordered_2[[16]]
#> [1] "x_mean"
#> 
#> $ordered_2[[17]]
#> [1] "x_AOC"
#> 
#> $ordered_2[[18]]
#> [1] "x_AOC_range"
#> 
#> $ordered_2[[19]]
#> [1] "xc50"
#> 
#> $ordered_2[[20]]
#> [1] "x_max"
#> 
#> $ordered_2[[21]]
#> [1] "ec50"
#> 
#> $ordered_2[[22]]
#> [1] "x_inf"
#> 
#> $ordered_2[[23]]
#> [1] "x_0"
#> 
#> $ordered_2[[24]]
#> [1] "h"
#> 
#> $ordered_2[[25]]
#> [1] "r2"
#> 
#> $ordered_2[[26]]
#> [1] "p_value"
#> 
#> $ordered_2[[27]]
#> [1] "rss"
#> 
#> $ordered_2[[28]]
#> [1] "x_sd_avg"
#> 
#> $ordered_2[[29]]
#> [1] "fit_type"
#> 
#> $ordered_2[[30]]
#> [1] "RV_mean"
#> 
#> $ordered_2[[31]]
#> [1] "GR_mean"
#> 
#> $ordered_2[[32]]
#> [1] "RV_AOC"
#> 
#> $ordered_2[[33]]
#> [1] "GR_AOC"
#> 
#> $ordered_2[[34]]
#> [1] "RV_AOC_range"
#> 
#> $ordered_2[[35]]
#> [1] "GR_AOC_range"
#> 
#> $ordered_2[[36]]
#> [1] "IC50"
#> 
#> $ordered_2[[37]]
#> [1] "GR50"
#> 
#> $ordered_2[[38]]
#> [1] "E_max"
#> 
#> $ordered_2[[39]]
#> [1] "GR_max"
#> 
#> $ordered_2[[40]]
#> [1] "EC50"
#> 
#> $ordered_2[[41]]
#> [1] "GEC50"
#> 
#> $ordered_2[[42]]
#> [1] "E_inf"
#> 
#> $ordered_2[[43]]
#> [1] "GR_inf"
#> 
#> $ordered_2[[44]]
#> [1] "E_0"
#> 
#> $ordered_2[[45]]
#> [1] "GR_0"
#> 
#> $ordered_2[[46]]
#> [1] "h_RV"
#> 
#> $ordered_2[[47]]
#> [1] "h_GR"
#> 
#> $ordered_2[[48]]
#> [1] "RV_r2"
#> 
#> $ordered_2[[49]]
#> [1] "GR_r2"
#> 
#> $ordered_2[[50]]
#> [1] "RV_p_value"
#> 
#> $ordered_2[[51]]
#> [1] "GR_p_value"
#> 
#> $ordered_2[[52]]
#> [1] "RV_rss"
#> 
#> $ordered_2[[53]]
#> [1] "GR_rss"
#> 
#> $ordered_2[[54]]
#> [1] "RV_sd_avg"
#> 
#> $ordered_2[[55]]
#> [1] "GR_sd_avg"
#> 
#> $ordered_2[[56]]
#> [1] "fit_type_RV"
#> 
#> $ordered_2[[57]]
#> [1] "fit_type_GR"
#> 
#> $ordered_2[[58]]
#> [1] "ReadoutValue"
#> 
#> $ordered_2[[59]]
#> [1] "BackgroundValue"
#> 
#> $ordered_2[[60]]
#> [1] "UntrtReadout"
#> 
#> $ordered_2[[61]]
#> [1] "Day0Readout"
#> 
#> $ordered_2[[62]]
#> [1] "masked"
#> 
#> $ordered_2[[63]]
#> [1] "ReferenceDivisionTime"
#> 
#> $ordered_2[[64]]
#> [1] "clid"
#> 
#> $ordered_2[[65]]
#> [1] "Gnumber"
#> 
#> $ordered_2[[66]]
#> [1] "Gnumber_2"
#> 
#> $ordered_2[[67]]
#> [1] "Gnumber_3"
#> 
#> $ordered_2[[68]]
#> [1] "Gnumber_4"
#> 
#> $ordered_2[[69]]
#> [1] "Gnumber_5"
#> 
#> $ordered_2[[70]]
#> [1] "Gnumber_6"
#> 
#> $ordered_2[[71]]
#> [1] "Gnumber_7"
#> 
#> $ordered_2[[72]]
#> [1] "Gnumber_8"
#> 
#> $ordered_2[[73]]
#> [1] "Gnumber_9"
#> 
#> $ordered_2[[74]]
#> [1] "Gnumber_10"
#> 
#> $ordered_2$barcode
#> [1] "Barcode" "Plate"  
#> 
#> $ordered_2$template
#> [1] "Template"  "Treatment"
#> 
#> $ordered_2$duration
#> [1] "Duration"
#> 
#> $ordered_2[[78]]
#> [1] "WellRow"
#> 
#> $ordered_2[[79]]
#> [1] "WellColumn"
#> 
#> 
#> $id
#> [1] "rId" "cId"
#> 
#> $iso_position
#> [1] "iso_level" "pos_x"     "pos_y"     "pos_x_ref" "pos_y_ref"
#> 
#> $excess
#> [1] "smooth"       "hsa_excess"   "bliss_excess"
#> 
#> $scores
#> [1] "hsa_score"   "bliss_score" "CIScore_50"  "CIScore_80" 
#> 
#> $isobolograms
#> [1] "normalization_type" "iso_level"          "pos_x"             
#> [4] "pos_y"              "pos_x_ref"          "pos_y_ref"         
#> [7] "log2_CI"            "log10_ratio_conc"  
#> 
#> $fit_source
#> [1] "fit_source"
#> 
#> $obsolete
#> [1] "RV"     "GR"     "Excess"
#> 
get_header("manifest")
#> $barcode
#> [1] "Barcode" "Plate"  
#> 
#> $template
#> [1] "Template"  "Treatment"
#> 
#> $duration
#> [1] "Duration"
#>