Merge assay data

Merge assay data

Usage

merge_assay(
  SElist,
  assay_name,
  additional_col_name = "data_source",
  discard_keys = NULL
)

Arguments

SElist

named list of Summarized Experiments

assay_name

name of the assay that should be extracted and merged

additional_col_name

string of column name that will be added to assay data for the distinction of possible duplicated metrics that can arise from multiple projects

discard_keys

character vector of string that will be discarded during creating BumpyMatrix object

Value

BumpyMatrix or list with data.table + BumpyMatrix

Examples

mae <- get_synthetic_data("finalMAE_combo_2dose_nonoise")

listSE <- list(
  combo1 = mae[[1]], 
  sa = mae[[2]]
)
merge_assay(listSE, "Normalized")
#> $DT
#>       data_source Concentration Concentration_2 masked normalization_type     x
#>            <char>         <num>           <num> <lgcl>             <fctr> <num>
#>    1:      combo1   0.001000000             0.0  FALSE                 RV 0.957
#>    2:      combo1   0.001000000             0.2  FALSE                 RV 0.957
#>    3:      combo1   0.001000000             1.0  FALSE                 RV 0.957
#>    4:      combo1   0.003162278             0.0  FALSE                 RV 0.712
#>    5:      combo1   0.003162278             0.2  FALSE                 RV 0.712
#>   ---                                                                          
#> 2804:          sa   1.000000000              NA  FALSE                 GR 1.000
#> 2805:          sa   1.000000000              NA  FALSE                 GR 1.000
#> 2806:          sa   1.000000000              NA  FALSE                 GR 1.000
#> 2807:          sa   1.000000000              NA  FALSE                 GR 1.000
#> 2808:          sa   1.000000000              NA  FALSE                 GR 1.000
#>       Gnumber DrugName drug_moa Gnumber_2 DrugName_2 drug_moa_2 Duration
#>        <char>   <char>   <char>    <char>     <char>     <char>    <num>
#>    1:  G00002 drug_002    moa_A    G00026   drug_026      moa_E       72
#>    2:  G00002 drug_002    moa_A    G00026   drug_026      moa_E       72
#>    3:  G00002 drug_002    moa_A    G00026   drug_026      moa_E       72
#>    4:  G00002 drug_002    moa_A    G00026   drug_026      moa_E       72
#>    5:  G00002 drug_002    moa_A    G00026   drug_026      moa_E       72
#>   ---                                                                   
#> 2804:  G00026 drug_026    moa_E      <NA>       <NA>       <NA>       72
#> 2805:  G00026 drug_026    moa_E      <NA>       <NA>       <NA>       72
#> 2806:  G00026 drug_026    moa_E      <NA>       <NA>       <NA>       72
#> 2807:  G00026 drug_026    moa_E      <NA>       <NA>       <NA>       72
#> 2808:  G00026 drug_026    moa_E      <NA>       <NA>       <NA>       72
#>          clid CellLineName   Tissue ReferenceDivisionTime
#>        <char>       <char>   <char>                 <num>
#>    1: CL00011  cellline_BA tissue_x                    26
#>    2: CL00011  cellline_BA tissue_x                    26
#>    3: CL00011  cellline_BA tissue_x                    26
#>    4: CL00011  cellline_BA tissue_x                    26
#>    5: CL00011  cellline_BA tissue_x                    26
#>   ---                                                    
#> 2804: CL00013  cellline_DA tissue_x                    34
#> 2805: CL00013  cellline_DA tissue_x                    34
#> 2806: CL00013  cellline_DA tissue_x                    34
#> 2807: CL00013  cellline_DA tissue_x                    34
#> 2808: CL00013  cellline_DA tissue_x                    34
#> 
#> $BM
#> 13 x 3 BumpyDataFrameMatrix
#> rownames: 1 2 ... 12 13 
#> colnames: 1 2 3 
#> preview [1,1]:
#>   DataFrame with 54 rows and 5 columns
#>          masked         x Concentration data_source normalization_type
#>       <logical> <numeric>     <numeric> <character>           <factor>
#>   1       FALSE    0.9570    0.00100000      combo1                 RV
#>   2       FALSE    0.4519    0.01000000      combo1                 GR
#>   3       FALSE    0.4519    0.01000000      combo1                 GR
#>   4       FALSE    0.7120    0.00316228      combo1                 RV
#>   5       FALSE    0.3691    0.03162278      combo1                 GR
#>   ...       ...       ...           ...         ...                ...
#>   50      FALSE    0.1863   10.00000000      combo1                 RV
#>   51      FALSE    0.7691    0.00316228      combo1                 GR
#>   52      FALSE    0.1866    3.16227766      combo1                 RV
#>   53      FALSE    0.1863   10.00000000      combo1                 RV
#>   54      FALSE    0.1866    3.16227766      combo1                 RV
#>